+++ src/validation/modules.rs
@@ -3,7 +3,7 @@ use crate::validation::instructions::{validate_constant_expression, validate_exp
 use crate::validation::types::{
     validate_function_type, validate_global_type, validate_memory_type, validate_table_type,
 };
-use crate::validation::validation::{Context, ValidationError, ValidationResult};
+use crate::validation::{Context, ValidationError, ValidationResult};
 
 // https://webassembly.github.io/spec/core/valid/modules.html#valid-func
 fn validate_function(context: &Context, function: &Function) -> ValidationResult {
